Move custom exceptions into subpackage
