+++ src/SmartWarehouse/CBC.java
@@ -4,7 +4,9 @@ import java.util.Deque;
 import java.util.concurrent.ConcurrentLinkedDeque;
 import java.util.concurrent.locks.Lock;
 import java.util.concurrent.locks.ReentrantLock;
-import src.SmartWarehouse.NoPacketException;
+
+import src.SmartWarehouse.exceptions.NoPacketException;
+import src.SmartWarehouse.exceptions.WrongPacketException;
 
 class CBC {
     private ConveyorBelt conveyorBelt = new ConveyorBelt();
