+++ src/SmartWarehouse/exceptions/NoPacketException.java
@@ -1,3 +1,3 @@
-package src.SmartWarehouse;
+package src.SmartWarehouse.exceptions;
 
 public class NoPacketException extends Exception {}
